pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey-Sherbakov",email = "63076931+Andrey-Sherbakov@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy (>=2.0.40)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false